Added RData
Added RData file without a few enormous objects. Before saving to Git, I removed the inla result object "out1" which was 800 MB, and a few other objects.  So you will need to rerun the model, and maybe the posterior sample. But lots of the other derived results are there.  @hrue