Draft 5 of manuscript and new code
@hrue ; @nicolelmichel;
Modified the model statement in the code to make it as simple as possible. Rewrote the code generally to make it more reproducible. Got rid of a bunch of data prep code and provided prepared input files instead. Still a little bit more simplification, archiving of useful but unecessarycode, to do before submission.

Rewrote the manuscript to reflect model changes. Redid the figures given new model formula and output. Added to the discussion.